[observe][android] Extract OTAnyValueSerializer to its own file
`OpenTelemetry.kt` was approaching 500 lines. Pull the custom
`KSerializer` implementation out so the main file is closer to plain
data-class declarations and OTel mappers.
